refactor: Remove extra abstraction layer on AuthService
