+++ integration-libs/cdc/src/auth/facade/cdc-auth.service.ts
@@ -3,10 +3,8 @@ import { Store } from '@ngrx/store';
 import {
   AsmAuthStorageService,
   AuthActions,
-  AuthService,
   AuthStorageService,
   AuthToken,
-  BasicAuthService,
   GlobalMessageService,
   GlobalMessageType,
   OCC_USER_ID_CURRENT,
@@ -22,17 +20,14 @@ import { CdcAuthActions } from '../store/actions';
 @Injectable({
   providedIn: 'root',
 })
-export class CdcAuthService extends AuthService {
+export class CdcAuthService {
   constructor(
     protected winRef: WindowRef,
     protected store: Store,
     protected authStorageService: AuthStorageService,
     protected userIdService: UserIdService,
-    protected basicAuthService: BasicAuthService,
     protected globalMessageService: GlobalMessageService
-  ) {
-    super(basicAuthService);
-  }
+  ) {}
 
   /**
    * Loads a new user token using custom oauth flow
@@ -124,24 +119,9 @@ export class CdcAuthService extends AuthService {
   }
 
   /**
-   * @override
-   *
-   * Logout a customer in storefront and CDC.
-   *
-   * @returns promise which resolves after completing logout
-   */
-  public logout(): Promise<any> {
-    return Promise.all([
-      super.logout(),
-      // trigger logout from CDC
-      this.logoutFromCdc(),
-    ]);
-  }
-
-  /**
    * Logout user from CDC
    */
-  protected logoutFromCdc(): void {
+  public logoutFromCdc(): void {
     this.winRef.nativeWindow?.['gigya']?.accounts?.logout();
   }
 }
