+++ projects/core/src/asm/services/asm-auth.service.ts
@@ -2,11 +2,11 @@ import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { combineLatest, from, Observable, of } from 'rxjs';
 import { map, switchMap, take } from 'rxjs/operators';
-import { AuthToken } from '../../auth';
 import { StateWithClientAuth } from '../../auth/client-auth/store/client-auth-state';
+import { AuthService } from '../../auth/user-auth/facade/auth.service';
 import { UserIdService } from '../../auth/user-auth/facade/user-id.service';
+import { AuthToken } from '../../auth/user-auth/models/auth-token.model';
 import { AuthRedirectService } from '../../auth/user-auth/services/auth-redirect.service';
-import { BasicAuthService } from '../../auth/user-auth/services/basic-auth.service';
 import { OAuthLibWrapperService } from '../../auth/user-auth/services/oauth-lib-wrapper.service';
 import { AuthActions } from '../../auth/user-auth/store/actions/index';
 import {
@@ -17,13 +17,13 @@ import { RoutingService } from '../../routing/facade/routing.service';
 import { AsmAuthStorageService, TokenTarget } from './asm-auth-storage.service';
 
 /**
- * Version of BasicAuthService that is working for both user na CS agent.
- * Overrides BasicAuthService when ASM module is enabled.
+ * Version of AuthService that is working for both user na CS agent.
+ * Overrides AuthService when ASM module is enabled.
  */
 @Injectable({
   providedIn: 'root',
 })
-export class AsmAuthService extends BasicAuthService {
+export class AsmAuthService extends AuthService {
   constructor(
     protected store: Store<StateWithClientAuth>,
     protected userIdService: UserIdService,
