+++ projects/core/src/auth/user-auth/services/index.ts
@@ -4,5 +4,4 @@ export * from './auth-redirect-storage.service';
 export * from './auth-redirect.service';
 export * from './auth-state-persistence.service';
 export * from './auth-storage.service';
-export * from './basic-auth.service';
 export * from './oauth-lib-wrapper.service';
